move Pub/Sub health autoconfiguration under main one
