+++ spring-cloud-gcp-pubsub/pom.xml
@@ -26,6 +26,11 @@
 			<artifactId>spring-integration-core</artifactId>
 			<optional>true</optional>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-actuator</artifactId>
+			<optional>true</optional>
+		</dependency>
 
 		<dependency>
 			<groupId>io.projectreactor</groupId>
